@@ -1,4 +1,4 @@
-namespace Domain.Enums
+namespace EquipmentRental.Common.Enums
 {
     public enum EquipmentType
     {
